+++ src/main/java/com/flab/foodeats/domain/menu/MenuOption.java
@@ -3,21 +3,21 @@ package com.flab.foodeats.domain.menu;
 
 public class MenuOption {
 
-	private int menuId;
-	private int menuOptionId;
+	private Long menuId;
+	private Long menuOptionId;
 	private String menuOptionName;
 	private int menuOptionPrice;
 
 	public MenuOption() {
 	}
 
-	public MenuOption(int menuId, String menuOptionName, int menuOptionPrice) {
+	public MenuOption(Long menuId, String menuOptionName, int menuOptionPrice) {
 		this.menuId = menuId;
 		this.menuOptionName = menuOptionName;
 		this.menuOptionPrice = menuOptionPrice;
 	}
 
-	public MenuOption(int menuId, int menuOptionId, String menuOptionName, int menuOptionPrice) {
+	public MenuOption(Long menuId, Long menuOptionId, String menuOptionName, int menuOptionPrice) {
 		this.menuId = menuId;
 		this.menuOptionId = menuOptionId;
 		this.menuOptionName = menuOptionName;
@@ -29,11 +29,11 @@ public class MenuOption {
 		this.menuOptionPrice = menuOptionPrice;
 	}
 
-	public int getMenuId() {
+	public Long getMenuId() {
 		return menuId;
 	}
 
-	public int getMenuOptionId() {
+	public Long getMenuOptionId() {
 		return menuOptionId;
 	}
